info_bot/pyproject.toml
[tool.poetry]
name = "info-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
asyncio = "^3.4.3"
requests = "^2.32.3"
httpx = "^0.28.1"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
